Integrated Global Network Graph  + Centrality into Index.py
